+++ SadConsole.Host.MonoGame/Renderers/ScreenSurfaceRenderer.cs
@@ -5,6 +5,7 @@ using SadRogue.Primitives;
 using Color = Microsoft.Xna.Framework.Color;
 using XnaRectangle = Microsoft.Xna.Framework.Rectangle;
 using SadConsole.Host.MonoGame;
+using System.Collections.Generic;
 
 namespace SadConsole.Renderers
 {
@@ -16,15 +17,22 @@ namespace SadConsole.Renderers
     /// </remarks>
     public class ScreenSurfaceRenderer : IRenderer
     {
-        /// <summary>
-        /// The blend state used by this renderer.
-        /// </summary>
-        public BlendState MonoGameBlendState { get; set; } = SadConsole.Host.Settings.MonoGameSurfaceBlendState;
+        protected IScreenSurface _screen;
 
         /// <summary>
         /// Color used with drawing the texture to the screen. Let's a surface become transparent.
         /// </summary>
-        protected Color _finalDrawColor = SadRogue.Primitives.Color.White.ToMonoColor();
+        public Color _finalDrawColor = SadRogue.Primitives.Color.White.ToMonoColor();
+
+        /// <summary>
+        /// Render steps to process.
+        /// </summary>
+        protected List<IRenderStep> RenderSteps = new List<IRenderStep>();
+
+        /// <summary>
+        /// The blend state used by this renderer.
+        /// </summary>
+        public BlendState MonoGameBlendState { get; set; } = SadConsole.Host.Settings.MonoGameSurfaceBlendState;
 
         /// <summary>
         /// Name of this renderer type.
@@ -45,14 +53,21 @@ namespace SadConsole.Renderers
         /// </summary>
         public RenderTarget2D BackingTexture;
 
+        /// <inheritdoc/>
+        public bool IsForced { get; set; }
+
         /// <summary>
         /// Cached set of rectangles used in rendering each cell.
         /// </summary>
-        protected XnaRectangle[] _renderRects;
+        public XnaRectangle[] CachedRenderRects;
 
         ///  <inheritdoc/>
         public virtual void Attach(IScreenSurface screen)
         {
+            _screen = screen;
+
+            for (int i = 0; i < RenderSteps.Count; i++)
+                RenderSteps[i].OnAdded(this, screen);
         }
 
         ///  <inheritdoc/>
@@ -60,62 +75,68 @@ namespace SadConsole.Renderers
         {
             BackingTexture?.Dispose();
             BackingTexture = null;
+
+            for (int i = 0; i < RenderSteps.Count; i++)
+                RenderSteps[i].OnRemoved(this, screen);
+
+            _screen = null;
         }
 
         ///  <inheritdoc/>
         public virtual void Render(IScreenSurface screen)
         {
+            for (int i = 0; i < RenderSteps.Count; i++)
+                RenderSteps[i].RenderStart();
+
             // If the tint is covering the whole area, don't draw anything
             if (screen.Tint.A != 255)
             {
                 // Draw call for surface
                 GameHost.Instance.DrawCalls.Enqueue(new DrawCalls.DrawCallTexture(BackingTexture, new Vector2(screen.AbsoluteArea.Position.X, screen.AbsoluteArea.Position.Y), _finalDrawColor));
-
-                // Draw call for cursors
-                foreach (var cursor in screen.GetSadComponents<Components.Cursor>())
-                {
-                    if (cursor.IsVisible && screen.Surface.IsValidCell(cursor.Position.X, cursor.Position.Y) && screen.Surface.View.Contains(cursor.Position))
-                    {
-                        GameHost.Instance.DrawCalls.Enqueue(
-                            new DrawCalls.DrawCallCell(cursor.CursorRenderCell,
-                                                        ((SadConsole.Host.GameTexture)screen.Font.Image).Texture,
-                                                        new XnaRectangle(screen.AbsoluteArea.Position.ToMonoPoint() + screen.Font.GetRenderRect(cursor.Position.X - screen.Surface.ViewPosition.X, cursor.Position.Y - screen.Surface.ViewPosition.Y, screen.FontSize).ToMonoRectangle().Location, screen.FontSize.ToMonoPoint()),
-                                                        screen.Font.SolidGlyphRectangle.ToMonoRectangle(),
-                                                        screen.Font.GetGlyphSourceRectangle(cursor.CursorRenderCell.Glyph).ToMonoRectangle()
-                                                        )
-                            );
-                    }
-                }
             }
+
+            for (int i = 0; i < RenderSteps.Count; i++)
+                RenderSteps[i].RenderBeforeTint();
 
             // If tint is visible, draw it
             if (screen.Tint.A != 0)
                 GameHost.Instance.DrawCalls.Enqueue(new DrawCalls.DrawCallColor(screen.Tint.ToMonoColor(), ((SadConsole.Host.GameTexture)screen.Font.Image).Texture, screen.AbsoluteArea.ToMonoRectangle(), screen.Font.SolidGlyphRectangle.ToMonoRectangle()));
+
+            for (int i = 0; i < RenderSteps.Count; i++)
+                RenderSteps[i].RenderEnd();
         }
 
         ///  <inheritdoc/>
         public virtual void Refresh(IScreenSurface screen, bool force = false)
         {
-            if (!force && !screen.IsDirty && BackingTexture != null) return;
+            IsForced = force;
 
             // Update texture if something is out of size.
             if (BackingTexture == null || screen.AbsoluteArea.Width != BackingTexture.Width || screen.AbsoluteArea.Height != BackingTexture.Height)
             {
                 BackingTexture?.Dispose();
                 BackingTexture = new RenderTarget2D(Host.Global.GraphicsDevice, screen.AbsoluteArea.Width, screen.AbsoluteArea.Height, false, Host.Global.GraphicsDevice.DisplayMode.Format, DepthFormat.Depth24);
+                IsForced = true;
             }
 
             // Update cached drawing rectangles if something is out of size.
-            if (_renderRects == null || _renderRects.Length != screen.Surface.View.Width * screen.Surface.View.Height || _renderRects[0].Width != screen.FontSize.X || _renderRects[0].Height != screen.FontSize.Y)
+            if (CachedRenderRects == null || CachedRenderRects.Length != screen.Surface.View.Width * screen.Surface.View.Height || CachedRenderRects[0].Width != screen.FontSize.X || CachedRenderRects[0].Height != screen.FontSize.Y)
             {
-                _renderRects = new XnaRectangle[screen.Surface.View.Width * screen.Surface.View.Height];
+                CachedRenderRects = new XnaRectangle[screen.Surface.View.Width * screen.Surface.View.Height];
 
-                for (int i = 0; i < _renderRects.Length; i++)
+                for (int i = 0; i < CachedRenderRects.Length; i++)
                 {
                     var position = SadRogue.Primitives.Point.FromIndex(i, screen.Surface.View.Width);
-                    _renderRects[i] = screen.Font.GetRenderRect(position.X, position.Y, screen.FontSize).ToMonoRectangle();
+                    CachedRenderRects[i] = screen.Font.GetRenderRect(position.X, position.Y, screen.FontSize).ToMonoRectangle();
                 }
+
+                IsForced = true;
             }
+
+            for (int i = 0; i < RenderSteps.Count; i++)
+                IsForced |= RenderSteps[i].RefreshPreStart();
+
+            if (!IsForced && !screen.IsDirty) return;
 
             // Render parts of the surface
             RefreshBegin(screen);
@@ -123,7 +144,13 @@ namespace SadConsole.Renderers
             if (screen.Tint.A != 255)
                 RefreshCells(screen.Surface, screen.Font);
 
+            for (int i = 0; i < RenderSteps.Count; i++)
+                RenderSteps[i].RefreshEnding();
+
             RefreshEnd(screen);
+
+            for (int i = 0; i < RenderSteps.Count; i++)
+                RenderSteps[i].RefreshEnd();
 
             screen.IsDirty = false;
         }
@@ -176,19 +203,50 @@ namespace SadConsole.Renderers
                     if (!cell.IsVisible) continue;
 
                     if (!cell.Background.Equals(Color.Transparent) && cell.Background != surface.DefaultBackground)
-                        Host.Global.SharedSpriteBatch.Draw(fontImage, _renderRects[rectIndex], font.SolidGlyphRectangle.ToMonoRectangle(), cell.Background.ToMonoColor(), 0f, Vector2.Zero, SpriteEffects.None, 0.3f);
+                        Host.Global.SharedSpriteBatch.Draw(fontImage, CachedRenderRects[rectIndex], font.SolidGlyphRectangle.ToMonoRectangle(), cell.Background.ToMonoColor(), 0f, Vector2.Zero, SpriteEffects.None, 0.3f);
 
                     if (!cell.Foreground.Equals(Color.Transparent))
-                        Host.Global.SharedSpriteBatch.Draw(fontImage, _renderRects[rectIndex], font.GetGlyphSourceRectangle(cell.Glyph).ToMonoRectangle(), cell.Foreground.ToMonoColor(), 0f, Vector2.Zero, cell.Mirror.ToMonoGame(), 0.4f);
+                        Host.Global.SharedSpriteBatch.Draw(fontImage, CachedRenderRects[rectIndex], font.GetGlyphSourceRectangle(cell.Glyph).ToMonoRectangle(), cell.Foreground.ToMonoColor(), 0f, Vector2.Zero, cell.Mirror.ToMonoGame(), 0.4f);
 
                     foreach (CellDecorator decorator in cell.Decorators)
                         if (!decorator.Color.Equals(Color.Transparent))
-                            Host.Global.SharedSpriteBatch.Draw(fontImage, _renderRects[rectIndex], font.GetGlyphSourceRectangle(decorator.Glyph).ToMonoRectangle(), decorator.Color.ToMonoColor(), 0f, Vector2.Zero, decorator.Mirror.ToMonoGame(), 0.5f);
+                            Host.Global.SharedSpriteBatch.Draw(fontImage, CachedRenderRects[rectIndex], font.GetGlyphSourceRectangle(decorator.Glyph).ToMonoRectangle(), decorator.Color.ToMonoColor(), 0f, Vector2.Zero, decorator.Mirror.ToMonoGame(), 0.5f);
 
                     i++;
                     rectIndex++;
                 }
             }
+        }
+
+        /// <inheritdoc/>
+        public void AddRenderStep(IRenderStep step)
+        {
+            RenderSteps.Add(step);
+            RenderSteps.Sort(CompareStep);
+
+            step.OnAdded(this, _screen);
+        }
+
+        /// <inheritdoc/>
+        public void RemoveRenderStep(IRenderStep step)
+        {
+            RenderSteps.Remove(step);
+            step.OnRemoved(this, _screen);
+        }
+
+        /// <inheritdoc/>
+        public IReadOnlyCollection<IRenderStep> GetRenderSteps() =>
+            RenderSteps.AsReadOnly();
+
+        private static int CompareStep(IRenderStep left, IRenderStep right)
+        {
+            if (left.SortOrder > right.SortOrder)
+                return 1;
+
+            if (left.SortOrder < right.SortOrder)
+                return -1;
+
+            return 0;
         }
 
         #region IDisposable Support
@@ -199,7 +257,7 @@ namespace SadConsole.Renderers
             if (!disposedValue)
             {
                 if (disposing)
-                    _renderRects = null;
+                    CachedRenderRects = null;
 
                 BackingTexture?.Dispose();
                 BackingTexture = null;